# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Email Sending Service",
    "version": "1.0.0",
    "description": "Unified gateway for transactional and broadcast email sending"
  },
  "servers": [
    {
      "url": "http://localhost:3009"
    }
  ],
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "service",
          "version"
        ]
      },
      "SendResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "messageId": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "enum": [
              "transactional",
              "broadcast"
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "provider"
        ]
      },
      "SendRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "transactional",
              "broadcast"
            ]
          },
          "appId": {
            "type": "string"
          },
          "clerkOrgId": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "campaignId": {
            "type": "string"
          },
          "clerkUserId": {
            "type": "string"
          },
          "to": {
            "type": "string",
            "format": "email"
          },
          "recipientFirstName": {
            "type": "string"
          },
          "recipientLastName": {
            "type": "string"
          },
          "recipientCompany": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "htmlBody": {
            "type": "string"
          },
          "textBody": {
            "type": "string"
          },
          "replyTo": {
            "type": "string",
            "format": "email"
          },
          "tag": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "appId",
          "clerkOrgId",
          "brandId",
          "campaignId",
          "to",
          "recipientFirstName",
          "recipientLastName",
          "recipientCompany",
          "subject"
        ]
      },
      "Stats": {
        "type": "object",
        "properties": {
          "sent": {
            "type": "number"
          },
          "delivered": {
            "type": "number"
          },
          "opened": {
            "type": "number"
          },
          "clicked": {
            "type": "number"
          },
          "replied": {
            "type": "number"
          },
          "bounced": {
            "type": "number"
          },
          "unsubscribed": {
            "type": "number"
          },
          "recipients": {
            "type": "number"
          }
        },
        "required": [
          "sent",
          "delivered",
          "opened",
          "clicked",
          "replied",
          "bounced",
          "unsubscribed",
          "recipients"
        ]
      },
      "StatsResponse": {
        "type": "object",
        "properties": {
          "transactional": {
            "$ref": "#/components/schemas/Stats"
          },
          "broadcast": {
            "$ref": "#/components/schemas/Stats"
          }
        }
      },
      "StatsRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "transactional",
              "broadcast"
            ]
          },
          "runIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "clerkOrgId": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "appId": {
            "type": "string"
          },
          "campaignId": {
            "type": "string"
          }
        }
      }
    },
    "parameters": {}
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/send": {
      "post": {
        "summary": "Send an email (transactional or broadcast)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "502": {
            "description": "Upstream service error"
          }
        }
      }
    },
    "/stats": {
      "post": {
        "summary": "Get aggregated email stats",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Aggregated stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "502": {
            "description": "Upstream service error"
          }
        }
      }
    },
    "/webhooks/postmark": {
      "post": {
        "summary": "Forward Postmark webhook events",
        "responses": {
          "200": {
            "description": "Webhook forwarded"
          },
          "502": {
            "description": "Upstream service error"
          }
        }
      }
    },
    "/webhooks/instantly": {
      "post": {
        "summary": "Forward Instantly webhook events",
        "responses": {
          "200": {
            "description": "Webhook forwarded"
          },
          "502": {
            "description": "Upstream service error"
          }
        }
      }
    }
  }
}